Added Intelligent Robber support for the AI
currently selects the tile that has the maximum amount of “other”
players sitting on it with settlements counting for 1 and cities for 2.
I looked at some stuff on the web about strategy and this is probably
even better than targetting 6 or 8s because people will naturally crowd
around the best possible spots and put their cities there first.