stored a saved best in self and check for a second move after building on a turn
best option stored as property "self.savedBestAction".  its a list with
index 0 as the key and 1 as the value from the option dictionary.
Initialzied to [None, None].  Will be like this until first decide move
call at which point will always be filled with a current unexecuted best
opt unless referred to in the middle of a decision or execution.  If
using in robber or trade funtions be mindful that it could be none if
the first roll of the game is a 7 or a trade is proposed before any
decisions have been made