Added Overall Scarcity and Centrality Function
Next is current scarcity on the map. the difference is overall is what
is available for all tiles. Current will rely on what players have
actually played next to. Ie the sum of all players expected income.
That value is probably the better basis for the short term of valuing
resources, but the overall scarcity is also interesting